// src/app/workflow-designer/models/enums.ts
export enum CursorType {
    ALIAS = 'alias',
    ALL_SCROLL = 'all-scroll',
    AUTO = 'auto',
    CELL = 'cell',
    CONTEXT_MENU = 'context-menu',
    COL_RESIZE = 'col-resize',
    COPY = 'copy',
    CROSSHAIR = 'crosshair',
    DEFAULT = 'default',
    E_RESIZE = 'e-resize',
    EW_RESIZE = 'ew-resize',
    GRAB = 'grab',
    GRABBING = 'grabbing',
    HELP = 'help',
    MOVE = 'move',
    N_RESIZE = 'n-resize',
    NE_RESIZE = 'ne-resize',
    NESW_RESIZE = 'nesw-resize',
    NS_RESIZE = 'ns-resize',
    NW_RESIZE = 'nw-resize',
    NWSE_RESIZE = 'nwse-resize',
    NO_DROP = 'no-drop',
    NONE = 'none',
    NOT_ALLOWED = 'not-allowed',
    POINTER = 'pointer',
    PROGRESS = 'progress',
    ROW_RESIZE = 'row-resize',
    S_RESIZE = 's-resize',
    SE_RESIZE = 'se-resize',
    SW_RESIZE = 'sw-resize',
    TEXT = 'text',
    URL = 'url',
    W_RESIZE = 'w-resize',
    WAIT = 'wait',
    ZOOM_IN = 'zoom-in',
    ZOOM_OUT = 'zoom-out'
}

export enum MouseButton {
    NONE, MAIN, AUXILURY, MIDDLE, BROWSER_BCK, BROWSER_FWD
}